arm64/configs: xenial: Do not build drivers as modules
There are a few reasons to do this, the most simple reasons are to avoid module mismatches.
In a lot of ROMs (including stock) you will not have working touchscreen, battery data and WiFi
if you're building kernel modules, this also helps for when TWRP is installed to the kernel ramdisk.

Signed-off-by: Lunarixus <Nathanb644@hotmail.co.uk>